context in a separate class
