+++ lib/gl/command.rb
@@ -7,3 +7,5 @@ require 'active_record'
 require 'rails/railtie'
 
 require 'gl/command/base'
+require 'gl/command/context'
+require 'gl/command/version'
